requirements.txt
accelerate==1.8.1
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
dnspython==2.7.0
einops==0.8.1
et_xmlfile==2.0.0
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.5.1
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.5
nltk==3.9.1
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
peft==0.16.0
pi==0.1.2
pillow==11.3.0
psutil==7.0.0
pymongo==4.13.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==5.0.0
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
